Click() method is unblocking async function now.
If any thread calls Click() and mouse was already processing Click() in any another (or in this) thread, the thread was blocked until the mouse button was released. Now if you call Click() and click is already in process, control flow will be returned to caller, but Click() will asynchronously (in distinct thread) wait for the mouse button to be released.